requirements.txt
pyjwt
python-dotenv
python-telegram-bot
python-telegram-bot[job-queue]
python-telegram-bot[rate-limiter]
pytz